virtual-tour-guide-IRWA-main/virtual-tour-guide-IRWA-main/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
better-profanity==0.7.0
cryptography==43.0.1
requests==2.32.3
